+++ tests/test_insights.py
@@ -119,14 +119,14 @@ class WorkerInsightsTest(unittest.TestCase):
         Insight containers are initially set to None values. When enabled they should be changed to appropriate
         containers. When a second task is started it should reset them. If disabled, they should remain None
         """
-        with WorkerPool(n_jobs=2) as pool:
+        with WorkerPool(n_jobs=2, enable_insights=True) as pool:
 
             # We run this a few times to see if it resets properly. We only verify this by checking the
             # n_completed_tasks
             for idx in range(3):
                 with self.subTest('enabled', idx=idx):
 
-                    pool.map(square, range(10), enable_insights=True, worker_init=self._init, worker_exit=self._exit)
+                    pool.map(square, range(10), worker_init=self._init, worker_exit=self._exit)
 
                     # Basic sanity checks for the values. Some max task args can be empty, in that case the duration
                     # should be 0 (= no data)
@@ -145,9 +145,11 @@ class WorkerInsightsTest(unittest.TestCase):
                             self.assertIn(args, {'Arg 0: 0', 'Arg 0: 1', 'Arg 0: 2', 'Arg 0: 3', 'Arg 0: 4',
                                                  'Arg 0: 5', 'Arg 0: 6', 'Arg 0: 7', 'Arg 0: 8', 'Arg 0: 9'})
 
+        with WorkerPool(n_jobs=2, enable_insights=False) as pool:
+
             # Disabling should set things to None again
             with self.subTest('disable'):
-                pool.map(square, range(10), enable_insights=False)
+                pool.map(square, range(10))
                 self.assertIsNone(pool._worker_insights.insights_manager)
                 self.assertIsNone(pool._worker_insights.insights_manager_lock)
                 self.assertIsNone(pool._worker_insights.worker_start_up_time)